+++ ZeroLibraries/Platform/Windows/Peripherals.cpp
@@ -47,7 +47,7 @@ bool GetGamepadState(size_t gamepadIndex, GamepadState* stateOut)
   if (!XInputModule || gamepadIndex >= cMaxGamepads)
     return false;
 
-  bool result = (*XInputGetState_P)(gamepadIndex, (XINPUT_STATE*)stateOut) == ERROR_SUCCESS;
+  bool result = (*XInputGetState_P)(static_cast<u32>(gamepadIndex), (XINPUT_STATE*)stateOut) == ERROR_SUCCESS;
 
   if (!result)
     return false;
@@ -81,7 +81,7 @@ bool SetGamepadVibration(size_t gamepadIndex, float leftSpeed, float rightSpeed)
   XINPUT_VIBRATION zerovibration;
   zerovibration.wLeftMotorSpeed = (WORD)(leftSpeed * MaxRumble);
   zerovibration.wRightMotorSpeed = (WORD)(rightSpeed * MaxRumble);
-  return (*XInputSetState_P)(gamepadIndex, &zerovibration) == ERROR_SUCCESS;
+  return (*XInputSetState_P)(static_cast<u32>(gamepadIndex), &zerovibration) == ERROR_SUCCESS;
 }
 
 }//namespace Zero
